rename Board to Game and give this class all exposed API methods that were previously in Game. Game renamed to GameManager
