--- src/action/move-action.ts.old
@@ -1,87 +1,0 @@
-import Action from './action';
-import type { Player } from '../player';
-
-import { GameElement, Piece } from '../board';
-
-import type {
-  Argument,
-  BoardSelection,
-  BoardQuerySingle,
-  SelectionDefinition
-} from './types';
-
-export default class MoveAction<P extends Player, E extends Piece<P>, S extends GameElement<P>> extends Action<P, []> {
-  piece?: BoardQuerySingle<P, E> | BoardSelection<P, E>
-  into?: BoardQuerySingle<P, E> | BoardSelection<P, E>
-  moveSelections: SelectionDefinition<P>[] = []
-
-  constructor(prompt: string) {
-    let movePiece: (...a: Argument<P>[]) => void;
-
-    if (noChoice(piece) && noChoice(to)) {
-      movePiece = () => resolve(piece)!.putInto(resolve(to)!);
-    }
-
-    if (noChoice(piece) && !noChoice(to)) {
-      selections = [{
-        prompt: promptTo || prompt,
-        selectOnBoard: {
-          chooseFrom: to.chooseFrom,
-        },
-        clientContext: { drag: piece }
-      }];
-      movePiece = (to: S) => resolve(piece)!.putInto(to);
-    }
-
-    if (!noChoice(piece) && noChoice(to)) {
-      selections = [{
-        prompt: prompt,
-        selectOnBoard: {
-          chooseFrom: piece.chooseFrom,
-        },
-        clientContext: { drop: to }
-      }];
-      movePiece = (piece: E) => piece.putInto(resolve(to, [piece])!);
-    }
-
-    if (!noChoice(piece) && !noChoice(to)) {
-      selections = [{
-        prompt: prompt,
-        selectOnBoard: {
-          chooseFrom: piece.chooseFrom,
-        },
-        clientContext: { drag: piece }
-      }, {
-        prompt: promptTo || prompt,
-        selectOnBoard: {
-          chooseFrom: to.chooseFrom,
-        },
-        clientContext: { drop: to }
-      }];
-      movePiece = (piece: E, to: S) => piece.putInto(to);
-    }
-
-    super({
-      prompt,
-      selections,
-      move: (...args: any[]) => {
-        movePiece(...args);
-        if (move) move(...args);
-      }
-    });
-  }
-
-  chooseInto<T extends GameElement<P>>(chooseFrom: BoardQueryMulti<P, T>, prompt?: string): Action<P, [...A, T]> {
-    this.selections.push(new Selection<P>({ prompt, selectOnBoard: { chooseFrom } }));
-    return this as unknown as Action<P, [...A, T]>;
-  }
-}
-
-const noChoice = <P extends Player>(s: BoardQuerySingle<P, GameElement<P>> | BoardSelection<P, GameElement<P>>): s is BoardQuerySingle<P, GameElement<P>> => {
-  return typeof s !== 'object' || !('chooseFrom' in s);
-}
-
-const resolve = <P extends Player, T extends GameElement<P>>(q: BoardQuerySingle<P, T>, ...args: Argument<P>[]) => {
-  if (typeof q === 'string') throw Error("not impl");
-  return (typeof q === 'function') ? q(...args) : q;
-}
