+++ src/board/index.ts
@@ -2,10 +2,10 @@ export { default as GameElement } from './element.js';
 export { default as Space } from './space.js';
 export { default as Piece } from './piece.js';
 export { default as Die } from './die.js';
-export { default as Board } from './board.js';
+export { default as Game } from './game.js';
 export { default as ElementCollection } from './element-collection.js';
 export { union } from './utils.js';
 
-export type { ActionLayout } from './board.js';
+export type { ActionLayout } from './game.js';
 export type { LayoutAttributes, Box, Vector } from './element.js';
 export type { ElementFinder, Sorter } from './element-collection.js';
