+++ src/board/piece.ts
@@ -1,14 +1,14 @@
 import GameElement from './element.js'
 import Space from './space.js'
 
-import type Board from './board.js'
+import type Game from './game.js'
 import type Player from '../player/player.js';
 
 /**
  * Pieces are game elements that can move during play
  * @category Board
  */
-export default class Piece<P extends Player<P, B> = any, B extends Board<P, B> = any> extends GameElement<P, B> {
+export default class Piece<P extends Player<P, B> = any, B extends Game<P, B> = any> extends GameElement<P, B> {
 
   /**
    * Move this piece into another element. This triggers any {@link
@@ -46,7 +46,7 @@ export default class Piece<P extends Player<P, B> = any, B extends Board<P, B> =
 
   /**
    * Remove this piece from the playing area and place it into {@link
-   * Board#pile}
+   * Game#pile}
    * @category Structure
    */
   remove() {
