+++ src/ui/lib.ts
@@ -5,15 +5,13 @@ import Selection from '../action/selection.js'
 import { GameElement, Die } from '../board/index.js'
 
 import type { GameStore } from './index.js';
-import type Game from '../game.js'
+import type { default as GameManager, SerializedMove, PendingMove } from '../game-manager.js'
 import type { Box, ElementJSON } from '../board/element.js'
 import type Player from '../player/player.js'
 import type { ResolvedSelection } from '../action/selection.js';
-import type { ActionLayout, Board, Piece } from '../board/index.js'
-import type { SerializedMove } from '../game.js'
-import type { PendingMove } from '../game.js'
+import type { ActionLayout, Game, Piece } from '../board/index.js'
 
-type GamePendingMoves = ReturnType<Game<Player, Board<Player>>['getPendingMoves']>;
+type GamePendingMoves = ReturnType<GameManager<Player, Game<Player>>['getPendingMoves']>;
 
 // this feels like the makings of a class
 export type UIMove = PendingMove<Player> & {
@@ -35,10 +33,10 @@ export type MoveMessage = {
 
 // refresh move and selections
 export function updateSelections(store: GameStore): GameStore {
-  let { game, position, move, placement } = store;
+  let { gameManager, position, move, placement } = store;
   if (!position) return store;
 
-  const player = game.players.atPosition(position);
+  const player = gameManager.players.atPosition(position);
   if (!player) return store;
   let state: GameStore = {
     ...store,
@@ -48,13 +46,13 @@ export function updateSelections(store: GameStore): GameStore {
   let maySubmit = !!move;
   let autoSubmit = false;
 
-  pendingMoves = game.getPendingMoves(player, move?.name, move?.args);
+  pendingMoves = gameManager.getPendingMoves(player, move?.name, move?.args);
 
   if (move && !pendingMoves?.moves) {
     // perhaps an update came while we were in the middle of a move
     console.error('move may no longer be valid. retrying getPendingMoves', move, pendingMoves);
     move = undefined;
-    pendingMoves = game.getPendingMoves(player);
+    pendingMoves = gameManager.getPendingMoves(player);
   }
 
   let moves = pendingMoves?.moves;
@@ -83,7 +81,7 @@ export function updateSelections(store: GameStore): GameStore {
 
       state = {
         ...state,
-        ...updateBoard(game, position),
+        ...updateBoard(gameManager, position),
       };
       // get the layout again since the updateBoard re-applied the layout after the piece was put into it
       const layout = into._ui.computedLayouts![layoutIndex];
@@ -112,7 +110,7 @@ export function updateSelections(store: GameStore): GameStore {
             ...moves[0],
             args: {...moves[0].args, [moves[0].selections[0].name]: arg}
           });
-          if (game.getPendingMoves(player, move.name, move.args)?.moves.length === 0) {
+          if (gameManager.getPendingMoves(player, move.name, move.args)?.moves.length === 0) {
             if (maySubmit === false) {
               autoSubmit = true;
               maySubmit = true;
@@ -127,7 +125,7 @@ export function updateSelections(store: GameStore): GameStore {
   // move is processable
   if (maySubmit && moves?.length === 0) {
     if (move) {
-      const player = game.players.atPosition(position);
+      const player = gameManager.players.atPosition(position);
       if (player) {
 
         // serialize now before we alter our state to ensure proper references
@@ -162,7 +160,7 @@ export function updateSelections(store: GameStore): GameStore {
           }
 
           try {
-            state.error = game.processMove({ player, ...move });
+            state.error = gameManager.processMove({ player, ...move });
 
             if (state.error) {
               // should probably never reach this point since the error would
@@ -170,23 +168,23 @@ export function updateSelections(store: GameStore): GameStore {
               throw Error(state.error);
             } else {
               try {
-                game.play();
+                gameManager.play();
               } catch (e) {
                 // this is a nice-to-have to speed up animations by
                 // "pre-playing" the flow. this could legitimately fail until
                 // the updated state is received from the server so exceptions
                 // here are ignored.
               }
-              game.sequence = Math.floor(game.sequence) + 0.5; // intermediate local update that will need to be merged
+              gameManager.sequence = Math.floor(gameManager.sequence) + 0.5; // intermediate local update that will need to be merged
               let json: any = undefined;
-              if (game.intermediateUpdates.length) {
-                json = game.intermediateUpdates[0][0].board
-                game.board.fromJSON(json);
-                game.intermediateUpdates = [];
+              if (gameManager.intermediateUpdates.length) {
+                json = gameManager.intermediateUpdates[0][0].board
+                gameManager.game.fromJSON(json);
+                gameManager.intermediateUpdates = [];
               }
               state = {
                 ...state,
-                ...updateBoard(game, position, json),
+                ...updateBoard(gameManager, position, json),
               }
 
               window.top!.postMessage(message, "*");
@@ -223,17 +221,17 @@ export function updateSelections(store: GameStore): GameStore {
     pendingMoves: pendingMoves?.moves as UIMove[],
   };
 
-  if (game.players.currentPosition.length > 0) {
-    const allowedActions = game.allowedActions(game.players.allCurrent()[0]);
+  if (gameManager.players.currentPosition.length > 0) {
+    const allowedActions = gameManager.allowedActions(gameManager.players.allCurrent()[0]);
     state.step = allowedActions.step;
     let description = allowedActions.description || 'taking their turn';
 
     const actionsWithDescription = allowedActions.actions.filter(a => a.description);
     if (actionsWithDescription.length === 1) {
       description = actionsWithDescription[0].description!;
-      if (!game.players.currentPosition.includes(position)) state.otherPlayerAction = actionsWithDescription[0].name;
+      if (!gameManager.players.currentPosition.includes(position)) state.otherPlayerAction = actionsWithDescription[0].name;
     }
-    state.actionDescription = `${game.players.currentPosition.length > 1 ? 'Players are' : game.players.current() + ' is'} ${description}`;
+    state.actionDescription = `${gameManager.players.currentPosition.length > 1 ? 'Players are' : gameManager.players.current() + ' is'} ${description}`;
   }
 
   state = {
@@ -259,9 +257,9 @@ export function updateSelections(store: GameStore): GameStore {
 // - a supplied layoutAction for the only current move
 // - a supplied layoutStep belonging to the step to which the current move(s) belong
 export function updateControls(store: GameStore): Pick<GameStore, "controls"> {
-  const { game, position, pendingMoves, selected, move, boardSelections, disambiguateElement, otherPlayerAction, step } = store;
+  const { gameManager, position, pendingMoves, selected, move, boardSelections, disambiguateElement, otherPlayerAction, step } = store;
 
-  if (!pendingMoves?.length && (!position || game.players.currentPosition.includes(position))) {
+  if (!pendingMoves?.length && (!position || gameManager.players.currentPosition.includes(position))) {
     return { controls: undefined };
   }
 
@@ -288,7 +286,7 @@ export function updateControls(store: GameStore): Pick<GameStore, "controls"> {
   // anchor to last element in arg list
   if (!layout && move && !pendingMoves?.[0].selections[0].isBoardChoice()) {
     const element = Object.entries(move.args).reverse().find(([name, el]) => (
-      !game.board._ui.stepLayouts["action:" + move.name]?.noAnchor?.includes(name) && el instanceof GameElement
+      !gameManager.game._ui.stepLayouts["action:" + move.name]?.noAnchor?.includes(name) && el instanceof GameElement
     ));
     if (element && (element[1] as GameElement)._ui?.computedStyle) {
       layout = { element: element[1] as GameElement, position: 'beside', gap: 2 };
@@ -302,7 +300,7 @@ export function updateControls(store: GameStore): Pick<GameStore, "controls"> {
     if (moves.length === 1) {
       // skip non-board moves if board elements already selected (cant this be more specific? just moves that could apply?)
       if (!selected.length || moves[0].selections.some(s => s.type !== 'board')) {
-        const actionLayout = game.board._ui.stepLayouts["action:" + moves[0].name];
+        const actionLayout = gameManager.game._ui.stepLayouts["action:" + moves[0].name];
         if (actionLayout?.element?._ui?.computedStyle) {
           layout = actionLayout;
           name = 'action:' + moves[0].name;
@@ -312,7 +310,7 @@ export function updateControls(store: GameStore): Pick<GameStore, "controls"> {
   }
 
   if (!layout && otherPlayerAction) {
-    const actionLayout = game.board._ui.stepLayouts["action:" + otherPlayerAction];
+    const actionLayout = gameManager.game._ui.stepLayouts["action:" + otherPlayerAction];
     if (actionLayout?.element?._ui?.computedStyle) {
       layout = actionLayout;
       name = 'action:' + otherPlayerAction;
@@ -321,12 +319,12 @@ export function updateControls(store: GameStore): Pick<GameStore, "controls"> {
 
   if (!layout && step) {
     name = 'step:' + step;
-    layout = game.board._ui.stepLayouts[name];
+    layout = gameManager.game._ui.stepLayouts[name];
   }
 
   if (!layout) {
     name = '*';
-    layout = game.board._ui.stepLayouts[name];
+    layout = gameManager.game._ui.stepLayouts[name];
   }
 
   if (layout) {
@@ -449,7 +447,7 @@ export function updateBoardPrompt(store: GameStore): Pick<GameStore, "boardPromp
   if (prompt) return { boardPrompt: prompt };
   if (actionDescription && !hasNonBoardMoves) return { boardPrompt: actionDescription };
   if (controls.moves.length && !hasNonBoardMoves) {
-    console.error(`No prompts defined for board actions (${controls.moves.map(m => m.name).join(', ')}). Add an action prompt or step prompt here.`);
+    console.error(`No prompts defined for actions (${controls.moves.map(m => m.name).join(', ')}). Add an action prompt or step prompt here.`);
     return {boardPrompt: '__missing__' };
   }
   return { boardPrompt: undefined };
@@ -461,17 +459,17 @@ export function removePlacementPiece(placement: Exclude<GameStore['placement'], 
 }
 
 // function to ensure react detects a change. must be called immediately after any function that alters board state
-export function updateBoard(game: Game<Player, Board<Player>>, position: number, json?: ElementJSON[]) {
+export function updateBoard(gameManager: GameManager<Player, Game<Player>>, position: number, json?: ElementJSON[]) {
   // rerun layouts. probably optimize TODO
-  game.contextualizeBoardToPlayer(game.players.atPosition(position));
-  game.board.applyLayouts(board => {
-    board.all(Die).appearance({
+  gameManager.contextualizeBoardToPlayer(gameManager.players.atPosition(position));
+  gameManager.game.applyLayouts(game => {
+    game.all(Die).appearance({
       render: (die: Die) => React.createElement(DieComponent, { die }),
       aspectRatio: 1,
     });
   });
 
-  return ({ boardJSON: json || game.board.allJSON() })
+  return ({ boardJSON: json || gameManager.game.allJSON() })
 }
 
 export function decorateUIMove(move: PendingMove<Player> | UIMove): UIMove {
